requirements.txt
flask
python-dotenv
youtube-transcript-api
